debug/elf: avoid using saferio to read data from SHT_NOBITS sections
A SHT_NOBITS section may have a nonzero size, but it occupies no space
in the file. The reader to read from this section is zeroReader, which
never reaches EOF. In this case, if the section size is incorrect and
holds a huge number, the saferio.ReadData does not help. What's worse,
it will eat up the memory a small block by a small block, and finally
results in an OOM.

This change calls "make([]byte, size)" to return the section data
directly. For an incorrect huge size, it panics with "runtime error:
makeslice: len out of range" fast before memory is allocated.

I think a further fix is to set a limit for the size of the SHT_NOBITS
section and refuse to read from this section when the size is larger
than the limit. But I can not find a documented limit for this section.

Fixes #54967.

Change-Id: I0c3ed4e097214fe88413d726a89122105ad45d4f
GitHub-Last-Rev: 51c6d7b0ed2cd9f44c528a85b3d074cd73304a04
GitHub-Pull-Request: golang/go#54994