debug/elf: return error on reading from SHT_NOBITS sections
An SHT_NOBITS section contains no bytes and occupies no space in the file.
This change makes it return an error on reading from this section so that
it will force the caller to check for an SHT_NNOBITS section.

We have considered another option to return "nil, nil" for the Data method.
It's abandoned because it might lead a program to simply do the wrong thing,
thinking that the section is empty.

Please note that it breaks programs which expect a byte slice with the
length described by the sh_size field. There are two reasons to introduce
this breaking change: 1. SHT_NOBITS means no data and it's unnecessary to
allocate memory for it; 2. it could reslut in an OOM if the file is
corrupted and has a huge sh_size.

Fixes #54967.

Change-Id: I0c3ed4e097214fe88413d726a89122105ad45d4f
GitHub-Last-Rev: 312d27d58a1fd25764704224e1dbb49f8917df3c
GitHub-Pull-Request: golang/go#54994