cmd/compile: if GOGC is not set, temporarily boost it for rapid starting heap growth
Benchmarking suggests about a 15% savings in user build time, about 4% savings
for wall time.  This helps most builds because small packages are common. See
https://perf.golang.org/search?q=upload:20221003.8 , for example. We've
talked about doing this in the past, the "new"(ish) metrics package makes it
a more tractable proposition.

The method here is:

1. If os.Getenv("GOGC") is empty, then increase GOGC to a  large value,
calculated to grow the heap to 32 + 4 * compile_parallelism before a GC
occurs (e.g., on a 4 processor box, 64M).  In practice, sometimes GC occurs
before that, but this still results in fewer GCs and saved time.

2. Use a finalizer to approximately detect when GC occurs, and use runtime
metrics to track progress towards the goal heap size, readjusting GOGC to
retarget it as necessary.  Reset GOGC to 100 when the heap is "close enough"
to the goal.

One feared failure mode of doing this is that the finalizer will be slow to
run and the heap will grow exceptionally large before GOGC is reset; I
monitored the heap size at reset across several boxes with a variety of
processor counts and extra noise (include several builds in parallel,
including a laptop with a busy many-tabs browser running) and the worst
overshoot was 50%.

I considered instead using the maximum heap size + GC-detecting-finalizer +
reset instead, but to me that seemed like it might have a worse bad-case
outcome; if the reset is delayed, it's possible the GC would start running
frequently, making it harder to run the finalizer, reach 50% utilization, and
the extra GCs would lose the advantage.  In practice, there are very few
boxes doing go builds that would be inconvenienced by the compiler briefly
consuming 128MB (a 100% overshoot, which was never observed) and the
processor-count dependence in the threshold is intended to hedge against
this.

This version of the CL includes a comment urging anyone who sees the code and
thinks it would work for them, to update a bug (to be created if CL is
approved) with information about their situation/experience, so that we may
consider creating some more official and reliable way of obtaining the same
result.

Change-Id: I45df1c927c1a7d7503ade1abd1a3300e27516633