testing: add an example showcasing B.RunParallel with B.ReportMetric
This commit was dedicated to adding an example of using B.ReportMetrics
with B.RunParallel called ExampleB_ReportMetric_parallel. In this
example, the same algorithm for ExampleB_ReportMetric was used, instead
with a concurrent for loop using PB.Next instead of a standard one.
There were also notes to describe best practice for the concurrent
looping and mention of the limitations of B.N.

Fixes: #50756
Change-Id: I2a621b4e367af5f4ec47d38a0da1035a8d52f628