os/exec: change the Interrupt field to a Cancel callback
As described in https://go.dev/issue/50436#issuecomment-1263591984.

The logic for correctly propagating Context errors upon cancellation
is somewhat subtle due to the inherent race between a command exiting
normally and a cancellation signal being delivered to it, so we should
not expect users to duplicate that logic in their own code or to start
their own goroutines to initiate cancellation asynchronously.

Because different commands can and do require different steps to
initiate clean shutdown (such as closing an input stream or sending a
request on a network socket), we should allow users to inject an
arbitrary callback instead of assuming that a Unix signal always
suffices.

Updates #50436.

Change-Id: I9f5ae29b55876e5861c2cc2c6fd779396a486a6f