cmd/go: add PGO auto mode
Add "auto" mode for the -pgo build flag. When -pgo=auto is
specified, if there is a default.pgo file in the directory of the
main package, it will be selected and used for the build.

A default.pgo file in a non-main package's directory is ignored.

When a go command invocation builds multiple main packages, the
profiles are ignored, as it doesn't know which one to use,
especially for a package that is depended from multiple main
packages (I think a go command invocation builds a package at most
ones?).

Do the same for "go install" and "go test". (Not for "go run" for
now as it is intended for a one-off build. (Should it?))

For #55022.

Change-Id: Iab7974ab8932daf0e83506de505e044a8e412466