net/http: refactor tests to run most in HTTP/1 and HTTP/2 modes
DO NOT REVIEW
DO NOT SUBMIT

Replace the ad-hoc approach to running tests in HTTP/1 and HTTP/2
modes with a 'run' function that executes a test in various modes.
By default, these modes are HTTP/1 and HTTP/2, but tests can
opt-in to HTTPS/1 as well.

The 'run' function also takes care of post-test cleanup (running the
afterTest function).

Update clientServerTest to use t.Cleanup to clean up after itself,
rather than leaving this up to tests to handle.

Change-Id: I16be64913c426fc93d84abc6ad85dbd3bc191224